Refactor QC logic for TS80
