+++ workspace/TS100/Core/Inc/QC3.h
@@ -7,7 +7,7 @@
 
 #ifndef INC_QC3_H_
 #define INC_QC3_H_
-
+#include "stdint.h"
 #ifdef __cplusplus
 extern "C" {
 #endif
